backend/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
librosa
soundfile
numpy
pillow
python-multipart
